+++ changelog.md
@@ -16,6 +16,7 @@ Codion Change Log
 - EntityTablePanel.Configuration.cellRenderer() now consumer based.
 - EntityTableExportPanel, attribute ordering implemented.
 - EntityTableExportModel, attribute ordering included in json.
+- EntityTableCellRendererFactory.create() now final, protected Swing specific create() added.
 
 ## 0.18.59
 ### is.codion.common.model